Tidying things up Rebuilding the starter kit

+++ ZauberCMS.Components/Editors/Models/TextContentSettingsModel.cs
@@ -2,5 +2,5 @@
 
 public class TextContentSettingsModel
 {
-    public int? MaxLength { get; set; }
+    public long? MaxLength { get; set; }
 }
